start of work to re-organize routers and handlers
