+++ spackbot/handlers/__init__.py
@@ -1,0 +1,4 @@
+from .pipelines import run_pipeline
+from .labels import add_labels
+from .reviewers import add_reviewers
+from .style import style_comment
